+++ CHANGELOG.md
@@ -172,6 +172,8 @@
   it is now referenced via the `Main` namespace instead of the project
   namespace,][6719] e.g. `Main.func1` instead of `MyProject.func1`. This makes
   it robust against project name changes.
+- [Added a button to return from an opened project back to the project
+  dashboard.][6474]
 
 [6279]: https://github.com/enso-org/enso/pull/6279
 [6421]: https://github.com/enso-org/enso/pull/6421
@@ -181,6 +183,7 @@
 [6663]: https://github.com/enso-org/enso/pull/6663
 [6752]: https://github.com/enso-org/enso/pull/6752
 [6719]: https://github.com/enso-org/enso/pull/6719
+[6474]: https://github.com/enso-org/enso/pull/6474
 
 #### EnsoGL (rendering engine)
 
@@ -781,6 +784,7 @@
 - [Limit number of reported warnings per value][6577]
 - [Suggestions are updated only when the type of the expression changes][6755]
 - [Add project creation time to project metadata][6780]
+- [Upgrade GraalVM to 22.3.1 JDK17][6750]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -891,6 +895,7 @@
 [6372]: https://github.com/enso-org/enso/pull/6372
 [6352]: https://github.com/enso-org/enso/pull/6352
 [6577]: https://github.com/enso-org/enso/pull/6577
+[6750]: https://github.com/enso-org/enso/pull/6750
 [6755]: https://github.com/enso-org/enso/pull/6755
 [6780]: https://github.com/enso-org/enso/pull/6780